+++ src/main/java/mainPackage/tmanager/services/UserService.java
@@ -25,7 +25,7 @@ public class UserService {
     }
 
 
-    public Optional<User> findById(int id) {
+    public Optional<User> findById(Long id) {
         return userRepository.findById(id);
     }
